Created VocalSystem type and rewrote Phone system
Now, Phones express themselves by messing with the tracks in a VocalSystem. Phones have complete control over which elements of the VocalSystem they utilize, and it is up to each Phone to figure out how to transition from the last phone.

Right now, I have created very basic Phone implementations. They probably will not result in nice-sounding speech, but it's a start. I tried to take some cues from the prototypes, but I did not struggle to make sure that the Phone system will produce everything the same way the prototypes do.