removed Amplitudes parameter from SawtoothTrack
I was not parameter anyway. Originally, the plan was to let the user set how much of each formant they want to use. Right now, however, it would be difficult to incorporate this notion into SawtoothParameters.powerFroFrequency(). In the future, I may revisit formant amplitude tuning if it seems necessary.